refactor: typed squads, CombatSimulator interface
