+++ src/main/java/unit/squad/CombatSimulator.java
@@ -1,0 +1,17 @@
+package unit.squad;
+
+import info.GameState;
+
+/**
+ * CombatSimulator interface for combat simulation and engagement evaluation.
+ */
+public interface CombatSimulator {
+
+    CombatResult evaluate(Squad squad, GameState gameState);
+
+    enum CombatResult {
+        ENGAGE,
+        RETREAT,
+        REGROUP
+    }
+}
